+++ src/dr_store/record_cache/__init__.py
@@ -2,6 +2,7 @@ from dr_store.record_cache.cache import (
     CacheEntry,
     CacheHit,
     RecordCache,
+    RecordCacheStats,
     derive_cache_key,
 )
 from dr_store.record_cache.sqlite import SqliteRecordCache
@@ -10,6 +11,7 @@ __all__ = [
     "CacheEntry",
     "CacheHit",
     "RecordCache",
+    "RecordCacheStats",
     "SqliteRecordCache",
     "derive_cache_key",
 ]